base feline atacado atualizada | bug de memória

+++ influir/bling/utils.bas
@@ -85,7 +85,7 @@ End Function
 
 Public Function all_unique(col_letter As String, sheet_name As String) As Variant
     Dim all_uniques(): all_uniques = Sheets(sheet_name).Range(col_letter & "6:" & col_letter & Sheets(sheet_name).Range("A1048576").End(xlUp).Row).Value
-    Dim dict_uniques As New Scripting.Dictionary, linha As Integer
+    Dim dict_uniques As New Scripting.Dictionary, linha As Long
     For linha = 1 To UBound(all_uniques)
         dict_uniques(all_uniques(linha, 1)) = Empty
     Next
